+++ src/main/java/com/cobber/fta/plugins/identity/SSN_CH.java
@@ -13,10 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cobber.fta.plugins;
+package com.cobber.fta.plugins.identity;
 
 import java.util.Locale;
 import java.util.Map;
+
+import org.apache.commons.validator.routines.checkdigit.CheckDigit;
+import org.apache.commons.validator.routines.checkdigit.CheckDigitException;
+import org.apache.commons.validator.routines.checkdigit.EAN13CheckDigit;
 
 import com.cobber.fta.AnalysisConfig;
 import com.cobber.fta.AnalyzerContext;
@@ -28,26 +32,32 @@ import com.cobber.fta.core.FTAType;
 import com.cobber.fta.token.TokenStreams;
 
 /**
- * Plugin to detect Japanese Individual Number (個人番号, kojin bangō), also known as My Number (マイナンバー, mai nambā).
+ * Plugin to detect Swiss SSN's (AVH Number/Sozialversicherungsnummer).
  */
-public class IdentityIN_JA extends LogicalTypeInfinite {
-	public static final String SEMANTIC_TYPE = "IDENTITY.INDIVIDUAL_NUMBER_JA";
-	private static final int IN_LENGTH = 12;
+public class SSN_CH extends LogicalTypeInfinite {
+	public static final String SEMANTIC_TYPE = "IDENTITY.SSN_CH";
+	private static final int SSN_LENGTH = 13;
 	public static final String BACKOUT_REGEXP = ".*";
 	private String regExp = BACKOUT_REGEXP;
+	private CheckDigit validator;
 
-	public IdentityIN_JA(final PluginDefinition plugin) {
+	public SSN_CH(final PluginDefinition plugin) {
 		super(plugin);
 	}
 
 	@Override
 	public boolean isCandidate(final String trimmed, final StringBuilder compressed, final int[] charCounts, final int[] lastIndex) {
-		if (trimmed.length() - charCounts[' '] != IN_LENGTH)
+		if (!trimmed.startsWith("756"))
+			return false;
+
+		if (trimmed.length() - (charCounts[' '] + charCounts['.']) != SSN_LENGTH)
 			return false;
 
 		for (int i = 0; i < trimmed.length(); i++) {
 			final char ch = trimmed.charAt(i);
-			if (ch != ' ' && !Character.isDigit(ch))
+			if (ch == ' ' || ch == '.')
+				continue;
+			if (!Character.isDigit(ch))
 				return false;
 		}
 
@@ -58,29 +68,29 @@ public class IdentityIN_JA extends LogicalTypeInfinite {
 	public boolean initialize(final Locale locale) throws FTAPluginException {
 		super.initialize(locale);
 
+		validator = new EAN13CheckDigit();
+
 		threshold = 98;
 
 		return true;
 	}
 
-	private int calculateCheckDigit(String in) {
-		long mySum = 0;
-		for (int i = 0; i < 5; i++)
-			mySum += (11 - i - 5) * (in.charAt(i) - '0');
-		for (int i = 5; i < in.length(); i++)
-			mySum += (11 - i + 1) * (in.charAt(i) - '0');
-
-		long check = 11 - mySum % 11;
-		if (check > 9)
-			check = 0;
-
-		return (int)check;
-	}
-
 	@Override
 	public String nextRandom() {
-		final String in = String.valueOf(random.nextInt(90000) + 10000) + String.format("%06d", random.nextInt(1000000));
-		return in + calculateCheckDigit(in);
+		final String avh = String.format("756%04d%04d%d",
+				random.nextInt(10000),
+				random.nextInt(10000),
+				random.nextInt(10));
+
+		String check = null;
+		try {
+			check = validator.calculate(avh);
+		} catch (CheckDigitException e) {
+			return null;
+		}
+
+		return avh.substring(0, 3) + "." + avh.substring(3, 7) + "." +
+			avh.substring(7, 11) + "." + avh.substring(11) + check;
 	}
 
 	@Override
@@ -105,20 +115,25 @@ public class IdentityIN_JA extends LogicalTypeInfinite {
 
 	@Override
 	public boolean isValid(final String input) {
-		final StringBuilder b = new StringBuilder(IN_LENGTH);
+		final String trimmed = input.trim();
+		if (!trimmed.startsWith("756"))
+			return false;
 
-		for (int i = 0; i < input.length(); i++) {
-			char ch = input.charAt(i);
-			if (ch == ' ')
+		final StringBuilder b = new StringBuilder(SSN_LENGTH);
+
+		for (int i = 0; i < trimmed.length(); i++) {
+			final char ch = input.charAt(i);
+			if (ch == ' ' || ch == '.')
 				continue;
 			if (!Character.isDigit(ch))
 				return false;
 			b.append(ch);
 		}
 
-		final int check = calculateCheckDigit(b.substring(0, IN_LENGTH - 1));
+		if (b.length() != SSN_LENGTH)
+			return false;
 
-		return '0' + check == b.charAt(11);
+		return validator.isValid(b.toString());
 	}
 
 	@Override